requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.18
django-dotenv==1.4.2
djangorestframework==3.15.2
idna==3.10
msal==1.31.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.2.3
